streamlit/requirements.txt
# dashboard/requirements.txt
streamlit>=1.10.0
kafka-python>=2.0.2
pandas>=1.3.3
plotly==5.15.0